ARM: run main() in a thread.
The crt0.S file now jumps directly to aica_init() after initialization of
the hardware. That function will create a task of the lowest priority
on which main() will run, if that function is available. If it is not,
a background task will be started, which just yield the thread indefinitely.